refactor: d.ts
