+++ app/components/guide/IGuide.ts
