+++ app/components/onionImage/IOnionImage.ts
