+++ app/components/ruler/IRuler.ts
